refactor: break up monolithic main.rs into modules (closes #133)
Splits main.rs (2,523 lines) into focused modules:
- src/cli.rs — Cli, Command, HooksAction structs
- src/output.rs — print/display formatting functions
- src/commands/mod.rs — shared helpers (load_and_discover, run_validation, etc.)
- src/commands/{check,coverage,generate,init,score,hooks,compact,resolve,diff,view,merge,issues}.rs

main.rs is now ~90 lines — just module declarations, main(), and run() dispatch.
All 80 tests pass, no behavioral changes.
